+++ src/components/FeaturePanel/FeatureHeading.tsx
@@ -3,8 +3,8 @@ import styled from 'styled-components';
 import EditIcon from '@mui/icons-material/Edit';
 import { IconButton } from '@mui/material';
 import { useEditDialogContext } from './helpers/EditDialogContext';
-import { PoiDescription } from './ImageSection/PoiDescription';
-import { StarButton } from './ImageSection/StarButton';
+import { PoiDescription } from './helpers/PoiDescription';
+import { StarButton } from './helpers/StarButton';
 import { getLabel } from '../../helpers/featureLabel';
 import { useFeatureContext } from '../utils/FeatureContext';
 import { t } from '../../services/intl';
