+++ src/components/FeaturePanel/helpers/StarButton.tsx
@@ -1,41 +1,20 @@
 import Star from '@mui/icons-material/Star';
 import StarBorder from '@mui/icons-material/StarBorder';
 import React from 'react';
-import { Tooltip } from '@mui/material';
+import { IconButton, Tooltip } from '@mui/material';
+import styled from 'styled-components';
 import { t } from '../../../services/intl';
 import { useStarsContext } from '../../utils/StarsContext';
-import { StyledActionButton } from './utils';
 import { useUserThemeContext } from '../../../helpers/theme';
 import { useIsClient } from '../../helpers';
 
-const StarButtonDarkPure = ({ isStarred, toggleStar }) => (
-  <StyledActionButton onClick={toggleStar}>
-    {isStarred ? (
-      <Tooltip arrow title={t('featurepanel.favorites_unsave_button')}>
-        <Star htmlColor="#fff" />
-      </Tooltip>
-    ) : (
-      <Tooltip arrow title={t('featurepanel.favorites_save_button')}>
-        <StarBorder htmlColor="#fff" />
-      </Tooltip>
-    )}
-  </StyledActionButton>
-);
-
-export const StarButtonDark = () => {
-  const { toggleStar, isStarred } = useStarsContext();
-  const isClient = useIsClient(); // SSR doesn't know localStorage, so it would end with hydrationWarning
-
-  return (
-    <>
-      {isClient ? (
-        <StarButtonDarkPure isStarred={isStarred} toggleStar={toggleStar} />
-      ) : (
-        <StarButtonDarkPure isStarred={false} toggleStar={() => {}} />
-      )}
-    </>
-  );
-};
+const StyledActionButton = styled(IconButton)<{ $color?: string }>`
+  svg {
+    width: 20px;
+    height: 20px;
+    color: ${({ $color }) => $color || '#fff'};
+  }
+`;
 
 const StarButtonPure = ({ isStarred, toggleStar }) => {
   const { currentTheme } = useUserThemeContext();
